+++ game/src/api/GameState.hpp
@@ -17,7 +17,7 @@ void fromString(const std::string &str);
         bool private_session;
         bool rolled;
         std::vector<double> rolls;
-        double turn_index;
+        int64_t turn_index;
         GameStateType type = static_cast<GameStateType>(0);
         std::vector<bool> used;
         bool victory;
