+++ game/src/api/IGameState.hpp
@@ -16,7 +16,7 @@ void fromString(const std::string &str);
         bool private_session;
         bool rolled;
         std::vector<double> rolls;
-        double turn_index;
+        int64_t turn_index;
         std::vector<bool> used;
         bool victory;
     };
