+++ game/src/api/ReconnectMsg.hpp
@@ -11,7 +11,7 @@ namespace API {
     struct ReconnectMsg {
 std::string toString() const;
 void fromString(const std::string &str);
-        double id;
+        int64_t id;
         ReconnectMsgType type = static_cast<ReconnectMsgType>(0);
     };
 }
