+++ game/src/api/UpdateNameMsg.hpp
@@ -11,7 +11,7 @@ namespace API {
     struct UpdateNameMsg {
 std::string toString() const;
 void fromString(const std::string &str);
-        double id;
+        int64_t id;
         std::string name;
         UpdateNameMsgType type = static_cast<UpdateNameMsgType>(0);
     };
